extracting solves reponsability from udp server
